test: clear cost rollup review blockers
